Response\HeadersTest: split `getValues()` specific tests to dedicated test class
